+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Please format the changes as follows:
 + Updates:
 
 # 1.0.32
++ BugFixes:
+  + Fix RPH bug - Allow SetEventMask post initialization if not affecting immutable flags
+  + Fix RPH bug - SetRejitMethodInfo with nullptr did not erase MethodInfo, causing AV
+  + Fix RPH bug - GetILFunctionBody allows NULL for out params if more than one
 
 # 1.0.31
 + BugFixes: